+++ decidim-admin/app/commands/decidim/admin/impersonate_managed_user.rb
@@ -25,7 +25,7 @@ module Decidim
         return broadcast(:invalid) if !user.managed? || !authorization_valid?
 
         create_impersonation_log
-        enque_expire_job
+        enqueue_expire_job
 
         broadcast(:ok)
       end
@@ -51,7 +51,7 @@ module Decidim
         )
       end
 
-      def enque_expire_job
+      def enqueue_expire_job
         Decidim::Admin::ExpireImpersonationJob
           .set(wait: Decidim::ImpersonationLog::SESSION_TIME_IN_MINUTES.minutes)
           .perform_later(user, current_user)